pyproject.toml
[project]
name = "ibkit"
version = "0.1.0"
description = "comparable companies analysis"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.119.0",
    "mysql-connector-python>=9.4.0",
    "numpy>=2.3.3",
    "pandas>=2.3.3",
    "plotly>=6.3.1",
    "pydantic>=2.12.0",
    "python-dotenv>=1.1.1",
    "redis>=6.4.0",
    "requests>=2.32.5",
    "scipy>=1.16.2",
    "uvicorn>=0.37.0",
    "yfinance>=0.2.66",
]

[tool.pylint.master]
init-hook = 'import sys; sys.path.append(".")'

[tool.pylint.messages_control]

[tool.pylint.format]
max-line-length = 100